Working both database and javalin communication
