+++ .idea/runConfigurations/Start.xml
@@ -1,6 +1,6 @@
 <component name="ProjectRunConfigurationManager">
   <configuration default="false" name="Start" type="Application" factoryName="Application">
-    <option name="MAIN_CLASS_NAME" value="ch.heigvd.dai.Main" />
+    <option name="MAIN_CLASS_NAME" value="ch.heigvd.dai.DatabaseTest" />
     <module name="MMO_Project" />
     <option name="PROGRAM_PARAMETERS" value="java -jar target/MMO_Project-1.0-SNAPSHOT.jar" />
     <method v="2">
